requirements.txt
# Requirements for RAG System

# Core LlamaIndex
llama-index>=0.9.0
llama-index-readers-file>=0.1.0
llama-index-readers-microsoft-onedrive>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-llms-ollama>=0.1.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.2

# Vector database
chromadb>=0.4.0

# Machine learning
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2

# Environment variables
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Optional: OneDrive support
onedrivesdk>=1.1.8

# Optional: Google Drive support
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.2
docx2txt>=0.8
